--- a/rust_con/.cargo/config.toml
+++ b/rust_con/.cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+rustflags = ["-Ctarget-cpu=native"]