on main with 60k posts
```
Benchmark 1: ./target/release/rust_rayon
Processing time (w/o IO): 134.500831ms
Processing time (w/o IO): 130.164506ms
Processing time (w/o IO): 125.846858ms
Processing time (w/o IO): 138.672469ms
Processing time (w/o IO): 129.986441ms
Processing time (w/o IO): 160.504282ms
Processing time (w/o IO): 138.690578ms
Processing time (w/o IO): 126.150532ms
Processing time (w/o IO): 148.820576ms
Processing time (w/o IO): 139.061629ms
Processing time (w/o IO): 137.511923ms
Processing time (w/o IO): 149.339401ms
Processing time (w/o IO): 130.010843ms
```

with this PR:
```
Benchmark 1: ./target/release/rust_rayon
Processing time (w/o IO): 115.480715ms
Processing time (w/o IO): 115.577198ms
Processing time (w/o IO): 105.337034ms
Processing time (w/o IO): 108.718906ms
Processing time (w/o IO): 105.843603ms
Processing time (w/o IO): 112.029415ms
Processing time (w/o IO): 115.705723ms
Processing time (w/o IO): 112.230522ms
Processing time (w/o IO): 112.120352ms
Processing time (w/o IO): 109.67582ms
Processing time (w/o IO): 119.478511ms
Processing time (w/o IO): 105.59841ms
Processing time (w/o IO): 109.165011ms
```