@jinyus 2023-11-05
Slower on VM...

Main (from readme):
`Rust Concurrent | 12.39 ms | 174.52 ms | 1.52s`

This PR:
```
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 15.698186ms
        Processing time (w/o IO): 15.828877ms
        Processing time (w/o IO): 15.550263ms
        Processing time (w/o IO): 15.630065ms
        Processing time (w/o IO): 15.469562ms
        Processing time (w/o IO): 15.429262ms
        Processing time (w/o IO): 15.388061ms
        Processing time (w/o IO): 15.334261ms
        Processing time (w/o IO): 15.518763ms
        Processing time (w/o IO): 15.363161ms
        Processing time (w/o IO): 15.477663ms
        Processing time (w/o IO): 15.614665ms
        Processing time (w/o IO): 15.578164ms
          Time (mean ± σ):      32.8 ms ±   0.5 ms    [User: 38.2 ms, System: 9.4 ms]
          Range (min … max):    32.0 ms …  33.5 ms    10 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 222.591877ms
        Processing time (w/o IO): 227.703164ms
        Processing time (w/o IO): 224.789015ms
          Time (mean ± σ):     289.4 ms ±   2.2 ms    [User: 486.7 ms, System: 25.6 ms]
          Range (min … max):   287.9 ms … 291.0 ms    2 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 1.982680548s
        Processing time (w/o IO): 1.977174655s
        Processing time (w/o IO): 1.976886576s
          Time (mean ± σ):      2.189 s ±  0.004 s    [User: 4.079 s, System: 0.071 s]
          Range (min … max):    2.186 s …  2.192 s    2 runs

```

Fresh run on main:
```
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 12.25691ms
        Processing time (w/o IO): 12.232909ms
        Processing time (w/o IO): 12.215509ms
        Processing time (w/o IO): 12.27991ms
        Processing time (w/o IO): 12.25161ms
        Processing time (w/o IO): 12.177308ms
        Processing time (w/o IO): 12.239409ms
        Processing time (w/o IO): 12.215709ms
        Processing time (w/o IO): 12.186809ms
        Processing time (w/o IO): 12.218909ms
        Processing time (w/o IO): 12.153508ms
        Processing time (w/o IO): 12.198809ms
        Processing time (w/o IO): 12.249209ms
          Time (mean ± σ):      30.3 ms ±   0.4 ms    [User: 31.0 ms, System: 11.2 ms]
          Range (min … max):    29.9 ms …  31.4 ms    10 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 175.998012ms
        Processing time (w/o IO): 174.464887ms
        Processing time (w/o IO): 175.767309ms
          Time (mean ± σ):     239.6 ms ±   1.4 ms    [User: 386.2 ms, System: 25.4 ms]
          Range (min … max):   238.7 ms … 240.6 ms    2 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 1.597206542s
        Processing time (w/o IO): 1.518112888s
        Processing time (w/o IO): 1.517240927s
          Time (mean ± σ):      1.729 s ±  0.000 s    [User: 3.153 s, System: 0.077 s]
          Range (min … max):    1.729 s …  1.729 s    2 runs
```